add safety margin to ticks' formula
Consider the following scenario:
where bar is the padding border with "uncertain" region,
1, 2, 3 are infectious people,
1 goes to the right,
2, 3 go to the left.

tick 0: 1|_2_3
tick 1: _|123_ with 1 moved to the right, 2 couldn't move bc of 1, and 3 moved
tick 2: _|123_ with no one moved, 3 stopped because cell to the left is occupied by 2.
Therefore, in tick 2 infection is going to spread from person_3 with affected position: 2 + infectionRadios (0-based positions). It is more than we previously considered by 1.